Unificando proyectos

+++ inc_config_letters.h
@@ -14,157 +14,157 @@ const int PROGMEM C_LET_A[] = {        O,I,O,EL,
                                        I,O,I,EA};
 
 
-const PROGMEM uint8_t C_LET_B[]={      I,I,O,EL,
+const PROGMEM int C_LET_B[]={      I,I,O,EL,
                                        I,O,I,EL,
                                        I,I,O,EL,
                                        I,O,I,EL,
                                        I,I,O,EA};
 
-const PROGMEM uint8_t C_LET_C[]={      I,I,I,EL,
+const PROGMEM int C_LET_C[]={      I,I,I,EL,
                                        I,O,O,EL,
                                        I,O,O,EL,
                                        I,O,O,EL,
                                        I,I,I,EA};
 
-const PROGMEM uint8_t C_LET_D[]={      I,I,O,EL,
+const PROGMEM int C_LET_D[]={      I,I,O,EL,
                                        I,O,I,EL,
                                        I,O,I,EL,
                                        I,O,I,EL,
                                        I,I,O,EA};
 
-const PROGMEM uint8_t C_LET_E[]={      I,I,I,EL,
+const PROGMEM int C_LET_E[]={      I,I,I,EL,
                                        I,O,O,EL,
                                        I,I,I,EL,
                                        I,O,O,EL,
                                        I,I,I,EA};
 
-const PROGMEM uint8_t C_LET_F[]={      I,I,I,EL,
+const PROGMEM int C_LET_F[]={      I,I,I,EL,
                                        I,O,O,EL,
                                        I,I,I,EL,
                                        I,O,O,EL,
                                        I,O,O,EA};
 
-const PROGMEM uint8_t C_LET_G[]={      O,I,I,EL,
+const PROGMEM int C_LET_G[]={      O,I,I,EL,
                                        I,O,O,EL,
                                        I,I,I,EL,
                                        I,O,I,EL,
                                        O,I,I,EA};
 
-const PROGMEM uint8_t C_LET_H[]={      I,O,I,EL,
+const PROGMEM int C_LET_H[]={      I,O,I,EL,
                                        I,O,I,EL,
                                        I,I,I,EL,
                                        I,O,I,EL,
                                        I,O,I,EA};
 
-const PROGMEM uint8_t C_LET_I[]={      O,I,O,EL,
+const PROGMEM int C_LET_I[]={      O,I,O,EL,
                                        O,I,O,EL,
                                        O,I,O,EL,
                                        O,I,O,EL,
                                        O,I,O,EA};
 
-const PROGMEM uint8_t C_LET_J[]={      O,O,I,EL,
+const PROGMEM int C_LET_J[]={      O,O,I,EL,
                                        O,O,I,EL,
                                        O,O,I,EL,
                                        O,O,I,EL,
                                        I,I,O,EA};
 
-const PROGMEM uint8_t C_LET_K[]={      I,O,I,EL,
+const PROGMEM int C_LET_K[]={      I,O,I,EL,
                                        I,O,I,EL,
                                        I,I,O,EL,
                                        I,O,I,EL,
                                        I,O,I,EA};
 
-const PROGMEM uint8_t C_LET_L[]={      I,O,O,EL,
+const PROGMEM int C_LET_L[]={      I,O,O,EL,
                                        I,O,O,EL,
                                        I,O,O,EL,
                                        I,O,O,EL,
                                        I,I,I,EA};
 
-const PROGMEM uint8_t C_LET_M[]={      I,O,I,EL,
+const PROGMEM int C_LET_M[]={      I,O,I,EL,
                                        I,I,I,EL,
                                        I,O,I,EL,
                                        I,O,I,EL,
                                        I,O,I,EA};
 
-const PROGMEM uint8_t C_LET_N[]={      I,O,I,EL,
+const PROGMEM int C_LET_N[]={      I,O,I,EL,
                                        I,O,I,EL,
                                        I,I,I,EL,
                                        I,I,I,EL,
                                        I,O,I,EA};
 
-const PROGMEM uint8_t C_LET_ENIE[]={   I,I,I,EL,
+const PROGMEM int C_LET_ENIE[]={   I,I,I,EL,
                                        O,O,O,EL,
                                        I,O,I,EL,
                                        I,I,I,EL,
                                        I,O,I,EA};
 
-const PROGMEM uint8_t C_LET_O[]={      O,I,O,EL,
+const PROGMEM int C_LET_O[]={      O,I,O,EL,
                                        I,O,I,EL,
                                        I,O,I,EL,
                                        I,O,I,EL,
                                        O,I,O,EA};
 
-const PROGMEM uint8_t C_LET_P[]={      I,I,O,EL,
+const PROGMEM int C_LET_P[]={      I,I,O,EL,
                                        I,O,I,EL,
                                        I,I,O,EL,
                                        I,O,O,EL,
                                        I,O,O,EA};
 
-const PROGMEM uint8_t C_LET_Q[]={      O,I,O,EL,
+const PROGMEM int C_LET_Q[]={      O,I,O,EL,
                                        I,O,I,EL,
                                        I,O,I,EL,
                                        I,I,I,EL,
                                        O,I,I,EA};
 
-const PROGMEM uint8_t C_LET_R[]={      I,I,O,EL,
+const PROGMEM int C_LET_R[]={      I,I,O,EL,
                                        I,O,I,EL,
                                        I,I,O,EL,
                                        I,O,I,EL,
                                        I,O,I,EA};
 
-const PROGMEM uint8_t C_LET_S[]={      O,I,I,EL,
+const PROGMEM int C_LET_S[]={      O,I,I,EL,
                                        I,O,O,EL,
                                        O,I,O,EL,
                                        O,O,I,EL,
                                        I,I,O,EA};
 
-const PROGMEM uint8_t C_LET_T[]={      I,I,I,EL,
+const PROGMEM int C_LET_T[]={      I,I,I,EL,
                                        O,I,O,EL,
                                        O,I,O,EL,
                                        O,I,O,EL,
                                        O,I,O,EA};
 
-const PROGMEM uint8_t C_LET_U[]={      I,O,I,EL,
+const PROGMEM int C_LET_U[]={      I,O,I,EL,
                                        I,O,I,EL,
                                        I,O,I,EL,
                                        I,O,I,EL,
                                        O,I,O,EA};
 
-const PROGMEM uint8_t C_LET_V[]={      I,O,I,EL,
+const PROGMEM int C_LET_V[]={      I,O,I,EL,
                                        I,O,I,EL,
                                        I,O,I,EL,
                                        O,I,O,EL,
                                        O,I,O,EA};
 
-const PROGMEM uint8_t C_LET_W[]={      I,O,I,EL,
+const PROGMEM int C_LET_W[]={      I,O,I,EL,
                                        I,O,I,EL,
                                        I,O,I,EL,
                                        I,I,I,EL,
                                        I,O,I,EA};
 
-const PROGMEM uint8_t C_LET_X[]={      I,O,I,EL,
+const PROGMEM int C_LET_X[]={      I,O,I,EL,
                                        I,O,I,EL,
                                        O,I,O,EL,
                                        I,O,I,EL,
                                        I,O,I,EA};
                                   
-const PROGMEM uint8_t C_LET_Y[]={      I,O,I,EL,
+const PROGMEM int C_LET_Y[]={      I,O,I,EL,
                                        I,O,I,EL,
                                        O,I,O,EL,
                                        O,I,O,EL,
                                        O,I,O,EA};
                                   
-const PROGMEM uint8_t C_LET_Z[]={      I,I,I,EL,
+const PROGMEM int C_LET_Z[]={      I,I,I,EL,
                                        O,O,I,EL,
                                        O,I,O,EL,
                                        I,O,O,EL,
@@ -172,163 +172,163 @@ const PROGMEM uint8_t C_LET_Z[]={      I,I,I,EL,
 
 // Big letters 4 pixel size 
 
-const PROGMEM uint8_t C_LET_BIG_A[]={  O,I,I,O,EL,
+const PROGMEM int C_LET_BIG_A[]={  O,I,I,O,EL,
                                        I,O,O,I,EL,
                                        I,I,I,I,EL,
                                        I,O,O,I,EL,
                                        I,O,O,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_B[]={  I,I,I,O,EL,
+const PROGMEM int C_LET_BIG_B[]={  I,I,I,O,EL,
                                        I,O,O,I,EL,
                                        I,I,I,O,EL,
                                        I,O,O,I,EL,
                                        I,I,I,O,EA};
 
-const PROGMEM uint8_t C_LET_BIG_C[]={  I,I,I,I,EL,
+const PROGMEM int C_LET_BIG_C[]={  I,I,I,I,EL,
                                        I,O,O,O,EL,
                                        I,O,O,O,EL,
                                        I,O,O,O,EL,
                                        I,I,I,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_D[]={  I,I,I,O,EL,
+const PROGMEM int C_LET_BIG_D[]={  I,I,I,O,EL,
                                        I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        I,I,I,O,EA};
 
-const PROGMEM uint8_t C_LET_BIG_E[]={  I,I,I,I,EL,
+const PROGMEM int C_LET_BIG_E[]={  I,I,I,I,EL,
                                        I,O,O,O,EL,
                                        I,I,I,I,EL,
                                        I,O,O,O,EL,
                                        I,I,I,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_F[]={  I,I,I,I,EL,
+const PROGMEM int C_LET_BIG_F[]={  I,I,I,I,EL,
                                        I,O,O,O,EL,
                                        I,I,I,I,EL,
                                        I,O,O,O,EL,
                                        I,O,O,O,EA};
 
-const PROGMEM uint8_t C_LET_BIG_G[]={  O,I,I,I,EL,
+const PROGMEM int C_LET_BIG_G[]={  O,I,I,I,EL,
                                        I,O,O,O,EL,
                                        I,O,I,I,EL,
                                        I,O,O,I,EL,
                                        O,I,O,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_H[]={  I,O,I,I,EL,
+const PROGMEM int C_LET_BIG_H[]={  I,O,I,I,EL,
                                        I,O,O,I,EL,
                                        I,I,I,I,EL,
                                        I,O,O,I,EL,
                                        I,O,O,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_I[]={  I,I,I,I,EL,
+const PROGMEM int C_LET_BIG_I[]={  I,I,I,I,EL,
                                        O,I,I,O,EL,
                                        O,I,I,O,EL,
                                        O,I,I,O,EL,
                                        I,I,I,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_J[]={  O,O,O,I,EL,
+const PROGMEM int C_LET_BIG_J[]={  O,O,O,I,EL,
                                        O,O,O,I,EL,
                                        O,O,O,I,EL,
                                        O,O,O,I,EL,
                                        I,I,I,O,EA};
 
-const PROGMEM uint8_t C_LET_BIG_K[]={  I,O,O,I,EL,
+const PROGMEM int C_LET_BIG_K[]={  I,O,O,I,EL,
                                        I,O,I,O,EL,
                                        I,I,O,O,EL,
                                        I,O,I,O,EL,
                                        I,O,O,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_L[]={  I,O,O,O,EL,
+const PROGMEM int C_LET_BIG_L[]={  I,O,O,O,EL,
                                        I,O,O,O,EL,
                                        I,O,O,O,EL,
                                        I,O,O,O,EL,
                                        I,I,I,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_M[]={  I,O,O,I,EL,
+const PROGMEM int C_LET_BIG_M[]={  I,O,O,I,EL,
                                        I,I,I,I,EL,
                                        I,I,I,I,EL,
                                        I,O,O,I,EL,
                                        I,O,O,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_N[]={  I,O,O,I,EL,
+const PROGMEM int C_LET_BIG_N[]={  I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        I,I,O,I,EL,
                                        I,O,I,I,EL,
                                        I,O,O,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_ENIE[]={ I,I,I,I,EL,
+const PROGMEM int C_LET_BIG_ENIE[]={ I,I,I,I,EL,
                                          O,O,O,O,EL,
                                          I,O,O,I,EL,
                                          I,I,O,I,EL,
                                          I,O,I,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_O[]={  O,I,I,O,EL,
+const PROGMEM int C_LET_BIG_O[]={  O,I,I,O,EL,
                                        I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        O,I,I,O,EA};
 
-const PROGMEM uint8_t C_LET_BIG_P[]={  I,I,I,O,EL,
+const PROGMEM int C_LET_BIG_P[]={  I,I,I,O,EL,
                                        I,O,O,I,EL,
                                        I,I,I,O,EL,
                                        I,O,O,O,EL,
                                        I,O,O,O,EA};
 
-const PROGMEM uint8_t C_LET_BIG_Q[]={  O,I,I,O,EL,
+const PROGMEM int C_LET_BIG_Q[]={  O,I,I,O,EL,
                                        I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        I,O,I,I,EL,
                                        O,I,I,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_R[]={  I,I,I,O,EL,
+const PROGMEM int C_LET_BIG_R[]={  I,I,I,O,EL,
                                        I,O,O,I,EL,
                                        I,I,I,O,EL,
                                        I,O,I,O,EL,
                                        I,O,O,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_S[]={  O,I,I,I,EL,
+const PROGMEM int C_LET_BIG_S[]={  O,I,I,I,EL,
                                        I,O,O,O,EL,
                                        O,I,I,O,EL,
                                        O,O,O,I,EL,
                                        I,I,I,O,EA};
 
-const PROGMEM uint8_t C_LET_BIG_T[]={  I,I,I,I,EL,
+const PROGMEM int C_LET_BIG_T[]={  I,I,I,I,EL,
                                        O,I,I,O,EL,
                                        O,I,I,O,EL,
                                        O,I,I,O,EL,
                                        O,I,I,O,EA};
 
-const PROGMEM uint8_t C_LET_BIG_U[]={  I,O,O,I,EL,
+const PROGMEM int C_LET_BIG_U[]={  I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        O,I,I,O,EA};
 
-const PROGMEM uint8_t C_LET_BIG_V[]={  I,O,O,I,EL,
+const PROGMEM int C_LET_BIG_V[]={  I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        O,I,I,O,EL,
                                        O,I,I,O,EA};
 
-const PROGMEM uint8_t C_LET_BIG_W[]={  I,O,O,I,EL,
+const PROGMEM int C_LET_BIG_W[]={  I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        I,I,I,I,EL,
                                        I,I,I,I,EL,
                                        I,O,O,I,EA};
 
-const PROGMEM uint8_t C_LET_BIG_X[]={  I,O,O,I,EL,
+const PROGMEM int C_LET_BIG_X[]={  I,O,O,I,EL,
                                        I,I,I,I,EL,
                                        O,I,I,O,EL,
                                        I,I,I,I,EL,
                                        I,O,O,I,EA};
                                   
-const PROGMEM uint8_t C_LET_BIG_Y[]={  I,O,O,I,EL,
+const PROGMEM int C_LET_BIG_Y[]={  I,O,O,I,EL,
                                        I,O,O,I,EL,
                                        O,I,I,O,EL,
                                        O,I,I,O,EL,
                                        O,I,I,O,EA};
                                   
-const PROGMEM uint8_t C_LET_BIG_Z[]={  I,I,I,I,EL,
+const PROGMEM int C_LET_BIG_Z[]={  I,I,I,I,EL,
                                        O,O,O,I,EL,
                                        O,I,I,O,EL,
                                        I,O,O,O,EL,
